move `server/handler/http/*` to `api/http/` (closes #1739) (#1758)
This removes the last of the dependencies from `api` on `server`, by moving `server/handler/http/error` and `server/handler/http/envelope` to `api/http`. It makes sense for these to be in `api`, because they codify how our API does error handling and object enveloping.

Nothing interesting in this one, it's all package renames.

fbshipit-source-id: 007bebc