server: observe 404 errors
This triggered a refactor where the "apiStats" field for each Route moved to become a field on the new RouteSet struct. Converting RouteSet from an array to a struct, triggered further changes.

stats gets a secret hook to allow stats to be emitted on stdout: this is useful for testing changes since it allows you to produce a list of stats being generated by a specific instance for comparison.

fbshipit-source-id: dfd36ff