[tbncollect] add `xds-only` command (#5079)
The point of this change is to add thexds-only command, which just serves xds, no collection.

along the way:

remove option to enable xds, it will always be on now
add an option to the stats package to make it ok to not have an API key for the API stats backend. If no api key is specified, a noop stats is returned.

fbshipit-source-id: 542a2c1